pyproject.toml
[tool.poetry]
name = "gendiff-hexlet-r3d3"
version = "0.1.0"
description = "Compares two configuration files and shows a difference."
authors = ["R3D3 <spoked-medals-0@icloud.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.run:main"

[tool.poetry.dependencies]
python = "^3.11"
pytest-cov = "^5.0.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
